Revert "Location and Weather services have separated from each other. OOP, Combine, Singleton"
