--- WeatherAround.xcodeproj/xcuserdata/misaperevozcikov.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "92A862C1-A1F7-4629-9EDA-EAEDA8FA33E9"
-   type = "1"
-   version = "2.0">
-</Bucket>
