+++ WeatherAround/Views/ViewModel/ViewModel.swift
@@ -6,17 +6,19 @@
 //
 
 import Foundation
+import CoreLocation
 import SwiftUI
-import Combine
 
 @MainActor class ViewModel: ObservableObject {
     
-    @ObservedObject var locationManager = LocationManager.shared
-    let weatherManager = WeatherManager.shared
+    let weatherService = WeatherService.shared
+//    let locationManager = LocationManager()
     
     @Published var weather: Weather?
     @Published var weatherHourly: WeatherHourly?
     @Published var weatherDaily: WeatherDaily?
+    
+    
     
     @Published var locationName = ""
     
@@ -27,28 +29,29 @@ import Combine
     
     @Published var loadingState: LoadingState = .loading
     
-    private (set) var cancellables = Set<AnyCancellable>()
-    
     
     init() {
-        fetchWeather()
-        weatherManager.setWeatherToModel { [weak self] current, daily in
+        //
+        //    ADD ALERTS TO ERRORS!
+        //
+        //
+        refresh()
+    }
+    
+    func refresh() {
+        weatherService.getLocation { [weak self] current, daily  in
             DispatchQueue.main.async {
-                
-                if let location = self?.locationManager.location {
-                    self?.weatherManager.getLocationName(for: location) { name in
-                        self?.locationName = name ?? "My location"
-                    }
-                } else {
-                    self?.loadingState = .failed
-                    print("Location not found")
-                }
                 
                 self?.weather = Weather(apiResponse: current)
                 self?.weatherHourly = WeatherHourly(apiResponse: current)
                 self?.weatherDaily = WeatherDaily(apiResponse: daily)
                 
-                
+            
+                if let location = self?.weatherService.location {
+                    self?.weatherService.getLocationName(for: location) { name in
+                        self?.locationName = name ?? "My location"
+                    }
+                }
                 withAnimation(.easeInOut) {
                     self?.loadingState = .loaded
                 }
@@ -56,21 +59,36 @@ import Combine
         }
     }
     
-    func fetchWeather() {
-        locationManager.$location
-            .sink { [weak self] location in
-                guard let location = location else {
-                    self?.loadingState = .failed
-                    print("Location not found")
-                    return
-                }
-                
-                Task  {
-                    await self?.weatherManager.downloadWeather(for: location)
-                }
-            }
-            .store(in: &cancellables)
-    }
+    
+//    func getWeather() {
+//
+//        guard let location = locationManager.location  else {
+//            print("There is no location")
+//            return
+//        }
+//        Task {
+//            await weatherService.downloadWeather(for: location) { [weak self] weatherResponse in
+//                DispatchQueue.main.async {
+//
+//                    self?.weather = Weather(apiResponse: weatherResponse)
+//                    self?.weatherHourly = WeatherHourly(apiResponse: weatherResponse)
+//
+//
+//                    if let location = self?.weatherService.location {
+//                        self?.weatherService.getLocationName(for: location) { name in
+//                            self?.locationName = name ?? "My location"
+//                        }
+//                    }
+//                    withAnimation(.easeInOut) {
+//                        self?.loadingState = .loaded
+//                    }
+//                }
+//
+//            }
+//        }
+//    }
+    
+    
 }
